# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/login": {
      "post": {
        "operationId": "AdminsController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario autenticado"
          },
          "403": {
            "description": "Acesso Negado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "summary": "Usuário Autenticado",
        "tags": [
          "Admins"
        ]
      }
    },
    "/admin/overview": {
      "get": {
        "operationId": "AdminsController_overview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "tags": [
          "Admins"
        ]
      }
    },
    "/admin": {
      "get": {
        "operationId": "AdminsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "tags": [
          "Admins"
        ]
      }
    },
    "/admin/{id}": {
      "get": {
        "operationId": "AdminsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "tags": [
          "Admins"
        ]
      },
      "patch": {
        "operationId": "AdminsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "tags": [
          "Admins"
        ]
      },
      "delete": {
        "operationId": "AdminsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        },
        "tags": [
          "Admins"
        ]
      }
    },
...
  },
  "info": {
    "title": "190Bahia",
    "description": "Projeto SSP-BA",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateAdminDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}